% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postinform.R
\name{postinform}
\alias{postinform}
\title{Process a batch of samples with Post-inForm.}
\usage{
postinform(
  input_file_or_dir,
  command = "all",
  immucan_output = FALSE,
  delete_input = FALSE
)
}
\arguments{
\item{input_file_or_dir}{String. Input file (.zip or .tar.gz) or directory to process.}

\item{command}{String [optional]. One of "check", "reduce" or "all". The default is "all".}

\item{immucan_output}{Boolean [optional]. If TRUE, the output is formated for the IMMUCAN}

\item{delete_input}{Boolean [optional]. If TRUE, the decompressed input data or the original
directory (if input_file_or_dir is passed a directory) is deleted when the function
completes successfully. Input .zip/.tar.gz files is never deleted.}
}
\value{
nothing.
}
\description{
Core function to run a Post-inForm analysis for a sample.
}
\details{
Core function to run a Post-inForm analysis for a sample.
}
\examples{
postinform(input_file_or_dir=input_file, delete_input=FALSE, immucan_output=FALSE)

}
